requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==2.4.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorama==0.4.6
comm==0.2.2
cssselect==1.3.0
debugpy==1.8.8
decorator==5.1.1
executing==2.1.0
fastavro==1.11.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.2
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
jieba3k==0.35.1
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.4.0
lxml_html_clean==0.4.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.6
packaging==24.2
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
